+++ src/app/(house)/search/page.tsx
@@ -1,10 +1,8 @@
 import { HydrationBoundary, dehydrate } from "@tanstack/react-query";
 import { GetSearchGoodListRequestDto } from "@/entities/goods/goods.types";
 import { searchGoodListInfiniteQueryOptions } from "@/features/goods/lib/searchGoodListInfiniteQueryOptions";
-import { FilterSection } from "@/widgets/goods/FilterSection";
-import { SearchGoodListSection } from "@/widgets/goods/SearchGoodListSection";
+import { SearchGoodListPage } from "@/pages/goods/ui/SearchGoodListPage";
 import { getQueryClient } from "@/shared/lib/react-query";
-import * as s from "@/app/(house)/search/style.css";
 
 type Props = {
   searchParams: Promise<
@@ -22,15 +20,8 @@ export default async function Search({ searchParams }: Props) {
   );
 
   return (
-    <main>
-      <div className={s.filterSection}>
-        <FilterSection />
-      </div>
-      <section className={s.searchGoodsSection}>
-        <HydrationBoundary state={dehydrate(queryClient)}>
-          <SearchGoodListSection q={q} order={order} />
-        </HydrationBoundary>
-      </section>
-    </main>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <SearchGoodListPage q={q} order={order} />
+    </HydrationBoundary>
   );
 }
