+++ src/pages/goods/ui/SearchGoodListPage/style.css.ts
@@ -17,7 +17,7 @@ export const filterSection = style([
   },
 ]);
 
-export const searchGoodsSection = style([
+export const searchGoodListSection = style([
   f.wFull,
   {
     margin: "0 auto",
@@ -35,4 +35,28 @@ export const searchGoodsSection = style([
       padding: `${ORDER_FILTER_HEIGHT}px 15px 15px`,
     },
   }),
+]);
+
+export const totalWrapper = style([
+  f.wFull,
+  {
+    marginBottom: "20px",
+  },
+]);
+
+export const total = style([
+  {
+    fontSize: "14px",
+    color: "#424242",
+  },
+]);
+
+export const emptyWrapper = style([
+  f.flex,
+  f.directionColumn,
+  f.alignCenter,
+  {
+    paddingBlock: "130px",
+    gap: "30px",
+  },
 ]);
